src/test/java/guru/springframework/sfgrecipes/services/IngredientServiceImplTest.java
package guru.springframework.sfgrecipes.services;

import guru.springframework.sfgrecipes.commands.IngredientCommand;
import guru.springframework.sfgrecipes.converters.IngredientToIngredientCommand;
import guru.springframework.sfgrecipes.converters.UnitOfMeasureToUnitOfMeasureCommand;
import guru.springframework.sfgrecipes.domain.Ingredient;
import guru.springframework.sfgrecipes.domain.Recipe;
import guru.springframework.sfgrecipes.repositories.RecipeRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.mockito.junit.jupiter.MockitoExtension;

import java.util.Optional;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.ArgumentMatchers.anyLong;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class IngredientServiceImplTest {

    public static final Long RECIPE_ID = 1L;
    public static final Long ING_ID = 2L;

    private IngredientToIngredientCommand ingredientToIngredientCommand;
    @Mock
    RecipeRepository recipeRepository;

    IngredientServiceImpl service;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
        this.ingredientToIngredientCommand = new IngredientToIngredientCommand(new UnitOfMeasureToUnitOfMeasureCommand());
        service = new IngredientServiceImpl(ingredientToIngredientCommand, recipeRepository);
    }

    @Test
    void testFindByRecipeIdAndId() {
        //given
        Recipe recipe = new Recipe().setId(RECIPE_ID);
        recipe.addIngredient(new Ingredient().setId(ING_ID));
        recipe.addIngredient(new Ingredient().setId(3L));
        Optional<Recipe> recipeOptional = Optional.of(recipe);
        when(recipeRepository.findById(anyLong())).thenReturn(recipeOptional);

        //when
        IngredientCommand command = service.findByRecipeIdAndIngredientId(RECIPE_ID, ING_ID);

        //then
        assertNotNull(command);
        assertEquals(command.getId(), ING_ID);
        assertEquals(command.getRecipeId(), RECIPE_ID);
    }
}